integrando cambios de selene funcionalidad de tarjeta

+++ grupo-tres-proyecto-final-java/src/main/java/models/Producto.java
@@ -12,7 +12,7 @@ public class Producto {
     @GenericGenerator(name = "native", strategy = "native")
     private long codigo_producto;
     private String nombre_producto;
-    private BigDecimal precio_producto;
+    private double precio_producto;
     private String descripcion_producto;
 
     @ManyToMany(mappedBy = "productos")
@@ -20,7 +20,7 @@ public class Producto {
 
     //Constructores.
     public Producto() { }
-    public Producto(String nombre_producto, BigDecimal precio_producto, String descripcion_producto) {
+    public Producto(String nombre_producto, double precio_producto, String descripcion_producto) {
         this.nombre_producto = nombre_producto;
         this.precio_producto = precio_producto;
         this.descripcion_producto = descripcion_producto;
@@ -30,14 +30,14 @@ public class Producto {
     //Getters.
     public long getCodigo_producto() { return codigo_producto; }
     public String getNombre_producto() { return nombre_producto; }
-    public BigDecimal getPrecio_producto() { return precio_producto; }
+    public double getPrecio_producto() { return precio_producto; }
     public String getDescripcion_producto() { return descripcion_producto; }
     public List<Compra> getCompras() { return compras; }
 
     //Setters.
     public void setCodigo_producto(long codigo_producto) { this.codigo_producto = codigo_producto; }
     public void setNombre_producto(String nombre_producto) { this.nombre_producto = nombre_producto; }
-    public void setPrecio_producto(BigDecimal precio_producto) { this.precio_producto = precio_producto; }
+    public void setPrecio_producto(double precio_producto) { this.precio_producto = precio_producto; }
     public void setDescripcion_producto(String descripcion_producto) { this.descripcion_producto = descripcion_producto; }
     public void setCompras(List<Compra> compras) { this.compras = compras; }
